# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API KosKuy PAA",
        "description": "Dokumentasi Swagger UI untuk testing endpoint API Laravel 12",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/room-types": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "List tipe kamar",
                "description": "Access: Semua user login. Menampilkan daftar tipe kamar beserta cabang, fasilitas, foto, dan jumlah kamar.",
                "operationId": "listRoomTypes",
                "parameters": [
                    {
                        "name": "branch_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Buat tipe kamar",
                "description": "Access: Admin dan Pemilik Kos. Membuat tipe kamar pada cabang tertentu, termasuk harga, ukuran, dan fasilitas kamar.",
                "operationId": "createRoomType",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_id",
                                    "name",
                                    "description",
                                    "price",
                                    "room_size"
                                ],
                                "properties": {
                                    "branch_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Kamar Standard"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Kamar nyaman dengan kamar mandi luar"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 750000
                                    },
                                    "room_size": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "facility_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/room-types/{room_type}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Detail tipe kamar",
                "description": "Access: Semua user login. Menampilkan detail tipe kamar, cabang, fasilitas, foto, dan daftar kamar.",
                "operationId": "showRoomType",
                "parameters": [
                    {
                        "name": "room_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update tipe kamar",
                "description": "Access: Admin dan Pemilik Kos. Mengubah informasi tipe kamar, harga, ukuran, status aktif, dan fasilitas kamar.",
                "operationId": "updateRoomType",
                "parameters": [
                    {
                        "name": "room_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "branch_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "room_size": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "facility_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Hapus tipe kamar",
                "description": "Access: Admin dan Pemilik Kos. Menghapus tipe kamar beserta kamar, foto, dan relasi fasilitasnya.",
                "operationId": "deleteRoomType",
                "parameters": [
                    {
                        "name": "room_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}